requirements.txt
dbt-core==1.7.11
dbt-clickhouse==1.7.5
airflow-clickhouse-plugin==1.2.0
astronomer-cosmos==1.3.2